manual: fix booking/uid page
