+++ apps/web/_app/_types.ts
@@ -1,0 +1,7 @@
+export type Params = {
+  [param: string]: string | string[] | undefined;
+};
+
+export type SearchParams = {
+  [param: string]: string | string[];
+};
